chore(api): move Country model to shared
