+++ api/tests/tooling/domain-builder/factory/build-country.js
@@ -1,4 +1,4 @@
-import { Country } from '../../../../src/certification/enrolment/domain/read-models/Country.js';
+import { Country } from '../../../../src/shared/domain/read-models/Country.js';
 import { normalizeAndSortChars } from '../../../../src/shared/infrastructure/utils/string-utils.js';
 
 const buildCountry = function ({ code = '99345', name = 'TOGO', matcher = normalizeAndSortChars(name) } = {}) {
